# docs/05-api-design/openapi.yaml
openapi: 3.1.0
info:
  title: Sibyl API
  version: "1.0"
  summary: Engineering Intelligence Platform — MVP API (PR Analysis, Test Impact Analysis, Flaky Detection, Root Cause Analysis)
  description: >
    Read-only analysis API plus the GitHub webhook ingestion endpoint. This API is
    the second surface after GitHub-native PR checks/comments (Stage 2 decision) —
    it does not re-expose data GitHub already serves (PR metadata, diffs); it only
    exposes Sibyl's derived analysis. All resources are addressed by owner/repo,
    matching GitHub's own convention; internal `installation_id` scoping (Stage 4)
    is resolved server-side and never appears in the public contract.
servers:
  - url: https://api.sibyl.dev/v1
    description: Placeholder production URL — finalized at Stage 7 (Infrastructure)
security:
  - bearerAuth: []
paths:
...
  /repositories/{owner}/{repo}/pulls/{prNumber}/regression-prediction:
    get:
      operationId: getRegressionPrediction
      summary: Regression Prediction result for a pull request (Sub-stage 9.10)
      tags: [regression-prediction]
      security:
        - bearerAuth: [read:regression-prediction]
      parameters:
        - $ref: "#/components/parameters/Owner"
        - $ref: "#/components/parameters/Repo"
        - $ref: "#/components/parameters/PrNumber"
      responses:
        "200":
          description: The latest predicted regression probability for this pull request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegressionPrediction"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT scoped per the RBAC catalog in `docs/05-api-design/README.md`
        (read:pr-analysis, read:test-intelligence, read:root-cause,
        read:dependency-analysis, read:regression-prediction,
        admin:installations, write:test-results, write:coverage-reports,
        write:dependency-manifests). Most scopes are issued after GitHub
        OAuth login; write:test-results, write:coverage-reports, and
        write:dependency-manifests are issued to CI jobs specifically (see
        the Stage 9.2, Stage 9.4, and Stage 9.6 addenda in
        `docs/05-api-design/README.md`).
  parameters:
    Owner:
      name: owner
      in: path
      required: true
      schema:
        type: string
    Repo:
      name: repo
      in: path
      required: true
      schema:
        type: string
    PrNumber:
      name: prNumber
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    NotFound:
      description: No such resource, or no analysis has ever been triggered for it.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem+json, with a Sibyl-specific `code` for programmatic handling.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable machine-readable error code, e.g. SIBYL_ANALYSIS_NOT_READY.
...
    RegressionPrediction:
      type: object
      required: [prNumber, headSha, regressionProbability, rationale, computedAt]
      properties:
        prNumber:
          type: integer
        headSha:
          type: string
        regressionProbability:
          type: number
          format: float
          minimum: 0
          maximum: 1
        rationale:
          type: string
        contributingSignals:
          type: array
          items:
            type: object
            properties:
              signal:
                type: string
              weight:
                type: number
        llmModel:
          type: string
        computedAt:
          type: string
          format: date-time
...